[Improve][Zeta] Move `RestJobExecutionEnvironment` to rest package
